refactor rpc db
