+++ host/src/bin/op-compose.rs
@@ -12,11 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{
-    collections::VecDeque,
-    fmt::Debug,
-    path::{Path, PathBuf},
-};
+// Example usage:
+//
+// RISC0_DEV_MODE=true RUST_LOG=info RUST_BACKTRACE=full ./target/release/op-compose \
+// --eth-rpc-url="https://eth-mainnet.g.alchemy.com/v2/API_KEY_HERE" \
+// --op-rpc-url="https://opt-mainnet.g.alchemy.com/v2/API_KEY_HERE" \
+// --cache \
+// --block-no=112875552 \
+// --blocks=8 \
+
+use std::{collections::VecDeque, fmt::Debug, path::PathBuf};
 
 use anyhow::Context;
 use clap::Parser;
@@ -25,19 +30,15 @@ use risc0_zkvm::{default_prover, serde::to_vec, Assumption, ExecutorEnv, Receipt
 use serde::{de::DeserializeOwned, Serialize};
 use zeth_guests::*;
 use zeth_lib::{
-    host::provider::{new_provider, BlockQuery},
+    host::rpc_db::RpcDb,
     optimism::{
-        batcher_db::{BatcherDb, BlockInput, MemDb},
+        batcher_db::BatcherDb,
         composition::{ComposeInput, ComposeInputOperation, ComposeOutputOperation},
         config::OPTIMISM_CHAIN_SPEC,
         DeriveInput, DeriveMachine,
     },
 };
-use zeth_primitives::{
-    block::Header,
-    transactions::{ethereum::EthereumTxEssence, optimism::OptimismTxEssence},
-    tree::MerkleMountainRange,
-};
+use zeth_primitives::{block::Header, tree::MerkleMountainRange};
 
 #[derive(Parser, Debug, Clone)]
 #[clap(author, version, about, long_about = None)]
@@ -67,25 +68,6 @@ struct Args {
     /// Runs the verification inside the zkvm executor locally. Accepts a custom maximum
     /// segment cycle count as a power of 2. [default: 20]
     local_exec: Option<u32>,
-}
-
-fn cache_file_path(cache_path: &Path, network: &str, block_no: u64, ext: &str) -> PathBuf {
-    cache_path
-        .join(network)
-        .join(block_no.to_string())
-        .with_extension(ext)
-}
-
-fn eth_cache_path(cache: &Option<PathBuf>, block_no: u64) -> Option<PathBuf> {
-    cache
-        .as_ref()
-        .map(|dir| cache_file_path(dir, "ethereum", block_no, "json.gz"))
-}
-
-fn op_cache_path(cache: &Option<PathBuf>, block_no: u64) -> Option<PathBuf> {
-    cache
-        .as_ref()
-        .map(|dir| cache_file_path(dir, "optimism", block_no, "json.gz"))
 }
 
 #[tokio::main]
@@ -376,135 +358,4 @@ pub fn prove(
 
     let prover = default_prover();
     prover.prove(env_builder.build().unwrap(), elf).unwrap()
-}
-
-#[derive(Clone)]
-
-pub struct RpcDb {
-    eth_rpc_url: Option<String>,
-    op_rpc_url: Option<String>,
-    cache: Option<PathBuf>,
-    mem_db: MemDb,
-}
-
-impl RpcDb {
-    pub fn new(
-        eth_rpc_url: Option<String>,
-        op_rpc_url: Option<String>,
-        cache: Option<PathBuf>,
-    ) -> Self {
-        RpcDb {
-            eth_rpc_url,
-            op_rpc_url,
-            cache,
-            mem_db: MemDb::new(),
-        }
-    }
-
-    pub fn get_mem_db(self) -> MemDb {
-        self.mem_db
-    }
-}
-
-impl BatcherDb for RpcDb {
-    fn get_full_op_block(
-        &mut self,
-        block_no: u64,
-    ) -> anyhow::Result<BlockInput<OptimismTxEssence>> {
-        let mut provider = new_provider(
-            op_cache_path(&self.cache, block_no),
-            self.op_rpc_url.clone(),
-        )?;
-        let block = {
-            let ethers_block = provider.get_full_block(&BlockQuery { block_no })?;
-            BlockInput {
-                block_header: ethers_block.clone().try_into().unwrap(),
-                transactions: ethers_block
-                    .transactions
-                    .into_iter()
-                    .map(|tx| tx.try_into().unwrap())
-                    .collect(),
-                receipts: None,
-            }
-        };
-        self.mem_db.full_op_block.insert(block_no, block.clone());
-        provider.save()?;
-        Ok(block)
-    }
-
-    fn get_op_block_header(&mut self, block_no: u64) -> anyhow::Result<Header> {
-        let mut provider = new_provider(
-            op_cache_path(&self.cache, block_no),
-            self.op_rpc_url.clone(),
-        )?;
-        let header: Header = provider
-            .get_partial_block(&BlockQuery { block_no })?
-            .try_into()?;
-        self.mem_db.op_block_header.insert(block_no, header.clone());
-        provider.save()?;
-        Ok(header)
-    }
-
-    fn get_full_eth_block(
-        &mut self,
-        block_no: u64,
-    ) -> anyhow::Result<BlockInput<EthereumTxEssence>> {
-        let query = BlockQuery { block_no };
-        let mut provider = new_provider(
-            eth_cache_path(&self.cache, block_no),
-            self.eth_rpc_url.clone(),
-        )?;
-        let block = {
-            let ethers_block = provider.get_full_block(&query)?;
-            let block_header: Header = ethers_block.clone().try_into().unwrap();
-            // include receipts when needed
-            let can_contain_deposits = zeth_lib::optimism::deposits::can_contain(
-                &OPTIMISM_CHAIN_SPEC.deposit_contract,
-                &block_header.logs_bloom,
-            );
-            let can_contain_config = zeth_lib::optimism::system_config::can_contain(
-                &OPTIMISM_CHAIN_SPEC.system_config_contract,
-                &block_header.logs_bloom,
-            );
-            let receipts = if can_contain_config || can_contain_deposits {
-                let receipts = provider.get_block_receipts(&query)?;
-                Some(
-                    receipts
-                        .into_iter()
-                        .map(|receipt| receipt.try_into())
-                        .collect::<anyhow::Result<Vec<_>, _>>()
-                        .context("invalid receipt")?,
-                )
-            } else {
-                None
-            };
-            BlockInput {
-                block_header,
-                transactions: ethers_block
-                    .transactions
-                    .into_iter()
-                    .map(|tx| tx.try_into().unwrap())
-                    .collect(),
-                receipts,
-            }
-        };
-        self.mem_db.full_eth_block.insert(block_no, block.clone());
-        provider.save()?;
-        Ok(block)
-    }
-
-    fn get_eth_block_header(&mut self, block_no: u64) -> anyhow::Result<Header> {
-        let mut provider = new_provider(
-            eth_cache_path(&self.cache, block_no),
-            self.eth_rpc_url.clone(),
-        )?;
-        let header: Header = provider
-            .get_partial_block(&BlockQuery { block_no })?
-            .try_into()?;
-        self.mem_db
-            .eth_block_header
-            .insert(block_no, header.clone());
-        provider.save()?;
-        Ok(header)
-    }
 }
